bugfixes and refactor
